api/requirements.txt
Flask==2.0.1
google-generativeai==0.1.0
python-dotenv==0.19.0
requests==2.28.1